*: fold ignored-dir status bench into worktree_status_bench_test.go
Keep all Status() benchmarks in a single file. Also refresh the
docstrings: they described the pre-PR behavior where ignored files were
walked and hashed, which no longer matches reality now that the
filesystem walker skips ignored directories.
